+++ doko/doko_simulator_test.go
@@ -1,150 +1,10 @@
 package doko
 
 import (
-	"strconv"
-	"strings"
 	"testing"
 
 	"github.com/scrouthtv/card-games/logic"
 )
-
-// GameStub is a game that is not connected to a hub or clients,
-// but instead saves no data besides the current state
-type GameStub struct {
-	state byte
-}
-
-// ID returns 5
-func (g *GameStub) ID() byte {
-	return 5
-}
-
-// Name returns " ## Invalid Test Game ## "
-func (g *GameStub) Name() string {
-	return " ## Invalid Test Game ## "
-}
-
-// State returns the current state
-func (g *GameStub) State() byte {
-	return g.state
-}
-
-// SetState sets the current state
-func (g *GameStub) SetState(state byte) {
-	g.state = state
-}
-
-// PlayerCount returns 4
-func (g *GameStub) PlayerCount() int {
-	return 4
-}
-
-// SendUpdates does nothing
-func (g *GameStub) SendUpdates() {
-
-}
-
-type DokoSim struct {
-	doko *Doko
-}
-
-func (ds *DokoSim) Move(move string) bool {
-	var p *logic.Packet = logic.NewPacket(move)
-
-	return ds.doko.PlayerMove(ds.doko.active, p)
-}
-
-func (ds *DokoSim) String() string {
-	var out strings.Builder
-	out.WriteString("Current Player: ")
-	out.WriteString(strconv.Itoa(ds.doko.active))
-
-	var i int
-	var deck *logic.Deck
-	for i = 0; i < len(ds.doko.hands); i++ {
-		deck = ds.doko.hands[i]
-		out.WriteString("\nHand ")
-		out.WriteString(strconv.Itoa(i))
-		out.WriteString(": ")
-		out.WriteString(deck.Short())
-	}
-
-	out.WriteString("\nTable: ")
-	out.WriteString(ds.doko.table.Short())
-
-	return out.String()
-}
-
-func (ds *DokoSim) TestHand(t *testing.T, player int, cards *logic.Deck) {
-	if t != nil {
-		t.Helper()
-	}
-	if !ds.doko.hands[player].Equal(cards) {
-		t.Errorf("Player %d has wrong hand:", player)
-		t.Logf("Expected: %s", cards)
-		t.Logf("Got: %s", ds.doko.hands[player])
-	}
-}
-
-func (ds *DokoSim) TestAllHands(t *testing.T, hands map[int]*logic.Deck) {
-	if t != nil {
-		t.Helper()
-	}
-	var player int
-	var deck *logic.Deck
-	for player, deck = range hands {
-		ds.TestHand(t, player, deck)
-	}
-}
-
-func (ds *DokoSim) TestWondeck(t *testing.T, player int, cards *logic.Deck) {
-	if t != nil {
-		t.Helper()
-	}
-	if !ds.doko.won[player].Equal(cards) {
-		t.Errorf("Player %d has wrong cards won:", player)
-		t.Logf("Expected: %s", cards)
-		t.Logf("Got: %s", ds.doko.won[player])
-	}
-}
-
-func (ds *DokoSim) TestAllWondecks(t *testing.T, won map[int]*logic.Deck) {
-	if t != nil {
-		t.Helper()
-	}
-	var player int
-	var deck *logic.Deck
-	for player, deck = range won {
-		ds.TestWondeck(t, player, deck)
-	}
-}
-
-func (ds *DokoSim) TestTable(t *testing.T, table *logic.Deck) {
-	if t != nil {
-		t.Helper()
-	}
-	if !ds.doko.table.Equal(table) {
-		t.Errorf("Table contents is wrong:")
-		t.Logf("Expected: %s", table)
-		t.Logf("Got: %s", ds.doko.table)
-	} else {
-		t.Logf("Table contents is as expected")
-	}
-}
-
-func (ds *DokoSim) TestProgress(t *testing.T, shouldp int, shouldt int) {
-	if t != nil {
-		t.Helper()
-	}
-
-	var isp, ist int = ds.doko.Progress()
-	if shouldp != -1 && isp != shouldp {
-		t.Errorf("Wrong progress, should be %d, is %d", shouldp, isp)
-	}
-	if shouldt != -1 && ist != shouldt {
-		t.Errorf("Wrong progress, should be %d, is %d", shouldt, ist)
-	}
-}
 
 // Example game:
 // hk ca c10 ca da h9 d9 s10 cq cj dj dk
@@ -426,46 +286,6 @@ func TestGameEnd(t *testing.T) {
 			t.Logf("Won %d: nil", i)
 		} else {
 			t.Logf("Won %d: %s", i, won.Short())
-		}
-	}
-}
-
-func (ds *DokoSim) addCardByShort(d *logic.Deck, short string) {
-	var c *logic.Card
-	var ok bool
-	ok, c = logic.CardFromShort(short)
-	if ok {
-		d.AddAll(c)
-	}
-}
-
-func (ds *DokoSim) assertCardMove(t *testing.T, short string, exp bool) {
-	if t != nil {
-		t.Helper()
-	}
-	var ok bool = ds.Move("card " + short)
-	if ok != exp {
-		if ok {
-			t.Error("Move did succeed, it shouldn't have")
-			t.FailNow()
-		} else {
-			t.Error("Move didn't succeed, it should have")
-			t.FailNow()
-		}
-	}
-}
-
-func (ds *DokoSim) assertPickup(t *testing.T, player int, exp bool) {
-	if t != nil {
-		t.Helper()
-	}
-	var p *logic.Packet = logic.NewPacket("pickup")
-	var ok bool = ds.doko.PlayerMove(player, p)
-	if ok != exp {
-		if ok {
-			t.Error("Pickup did succeed, it shouldn't have")
-		} else {
-			t.Error("Pickup didn't succeed, it should have")
 		}
 	}
 }
